domain-info/requirements.txt
Flask
python-whois
